+++ lib/Frontend/ParseableInterfaceModuleLoader.cpp
@@ -26,7 +26,6 @@
 #include "swift/Frontend/ParseableInterfaceSupport.h"
 #include "swift/Frontend/PrintingDiagnosticConsumer.h"
 #include "swift/SILOptimizer/PassManager/Passes.h"
-#include "swift/Serialization/ModuleFormat.h"
 #include "swift/Serialization/SerializationOptions.h"
 #include "swift/Serialization/Validation.h"
 #include "clang/Basic/Module.h"
